Use 16-bit addresses, not 32

+++ bus.go
@@ -16,14 +16,14 @@ func (b *Bus) Reset() {
 	b.mmap.Reset()
 }
 
-func (b *Bus) Read8(addr uint32) uint8 {
+func (b *Bus) Read8(addr uint16) uint8 {
 	return b.mmap.Read8(addr)
 }
 
-func (b *Bus) Write8(addr uint32, val uint8) {
+func (b *Bus) Write8(addr uint16, val uint8) {
 	b.mmap.Write8(addr, val)
 }
 
-func (b *Bus) MapSlice(addr, end uint32, buf []byte) {
+func (b *Bus) MapSlice(addr, end uint16, buf []byte) {
 	b.mmap.MapSlice(addr, end, buf)
 }
